rm getProfile from protocol, use ProfileDetails
